# dashboard/.schema-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Orchestrator Dashboard",
        "description": "REST/SSE API of the YAOE-FLOW observability dashboard.",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/settings/{key}": {
            "put": {
                "operationId": "putApiSettingsByKey",
                "tags": [
                    "Settings"
                ],
                "summary": "Grava setting no banco",
                "responses": {
                    "200": {
                        "description": "Salvo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "const": true
                                        },
                                        "groups": {
                                            "type": "object",
                                            "propertyNames": {
                                                "type": "string"
                                            },
                                            "additionalProperties": {}
                                        }
                                    },
                                    "required": [
                                        "ok",
                                        "groups"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de valida\u00e7\u00e3o",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "key",
                        "schema": {
                            "type": "string",
                            "minLength": 1
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "value": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "value"
                                ]
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "deleteApiSettingsByKey",
                "tags": [
                    "Settings"
                ],
                "summary": "Reseta setting pro default (remove do banco)",
                "responses": {
                    "200": {
                        "description": "Resetado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "const": true
                                        },
                                        "groups": {
                                            "type": "object",
                                            "propertyNames": {
                                                "type": "string"
                                            },
                                            "additionalProperties": {}
                                        }
                                    },
                                    "required": [
                                        "ok",
                                        "groups"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "key",
                        "schema": {
                            "type": "string",
                            "minLength": 1
                        },
                        "required": true
                    }
                ]
            }
        },
        "/api/settings/validate-linear": {
            "post": {
                "operationId": "postApiSettingsValidateLinear",
                "tags": [
                    "Settings"
                ],
                "summary": "Valida STATE_*/LABEL_* contra workflow states e labels do Linear",
                "responses": {
                    "200": {
                        "description": "Resultado da valida\u00e7\u00e3o",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "propertyNames": {
                                        "type": "string"
                                    },
                                    "additionalProperties": {}
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Erro de conex\u00e3o com Linear",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}